sched: fair: balance for single core cluster
Android will unset SD_LOAD_BALANCE for single core cluster domain and
for some product it is true to have a single core cluster and the MC
domain thus lacks the SD_LOAD_BALANCE flag. This will cause
select_task_rq_fair logic for that core. And the task will spin forever
in that core.

Bug: 141334320
Test: boot and see task on core7 scheduled correctly
Change-Id: I7c2845b1f7bc1d4051eb3ad6a5f9838fb0b1ba04
Signed-off-by: Wei Wang <wvw@google.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>